requirements.txt
langgraph>=1.0.0
langchain-openai>=1.0.0
langchain-core>=0.3.0
tavily-python>=0.5.0
requests
pydantic>=2.0
python-dotenv
